Refactor intro into IntroScene
