+++ src/openrct2-ui/drawing/engines/opengl/OpenGLDrawingEngine.cpp
@@ -25,7 +25,6 @@
 #    include <algorithm>
 #    include <cmath>
 #    include <openrct2-ui/interface/Window.h>
-#    include <openrct2/Intro.h>
 #    include <openrct2/config/Config.h>
 #    include <openrct2/core/Console.hpp>
 #    include <openrct2/drawing/Drawing.h>
