+++ src/openrct2-ui/windows/TitleExit.cpp
@@ -10,7 +10,6 @@
 #include <openrct2-ui/interface/Widget.h>
 #include <openrct2-ui/windows/Window.h>
 #include <openrct2/Context.h>
-#include <openrct2/Intro.h>
 #include <openrct2/config/Config.h>
 #include <openrct2/localisation/Localisation.h>
 #include <openrct2/sprites.h>
@@ -38,9 +37,6 @@ static Widget _titleExitWidgets[] = {
 
         void OnMouseUp(WidgetIndex widgetIndex) override
         {
-            if (gIntroState != IntroState::None)
-                return;
-
             switch (widgetIndex)
             {
                 case WIDX_EXIT:
