+++ src/openrct2/paint/Painter.cpp
@@ -11,7 +11,6 @@
 
 #include "../Game.h"
 #include "../GameState.h"
-#include "../Intro.h"
 #include "../OpenRCT2.h"
 #include "../ReplayManager.h"
 #include "../config/Config.h"
@@ -24,6 +23,7 @@
 #include "../localisation/Language.h"
 #include "../paint/Paint.h"
 #include "../profiling/Profiling.h"
+#include "../scenes/intro/IntroScene.h"
 #include "../scenes/title/TitleScene.h"
 #include "../ui/UiContext.h"
 #include "../world/TileInspector.h"
@@ -43,7 +43,8 @@ void Painter::Paint(IDrawingEngine& de)
     PROFILED_FUNCTION();
 
     auto dpi = de.GetDrawingPixelInfo();
-    if (gIntroState != IntroState::None)
+
+    if (IntroIsPlaying())
     {
         IntroDraw(*dpi);
     }
